Refactor component logic into Person
The `Component` enum was previously duplicated in the respective
predicate types. The predicates handled the component extraction from
`Person` and were needed to be handled differently by the
`FilterCommandParser`.

To fix this, the `Component` enum has been moved to `Person` along with
the extraction logic. This removes the violation of the Law of Demeter
from the extraction code and removes the dependency of the
`FilterCommandParser` from every type to just `Person`.

`ComponentExistencePredicate`'s constructor now throws an error if given
an unsupported component to accomodate the generalising of the
`Component` code. This will also make error reporting on unsupported
components more cohesive as every class handles their own error.